side panel viewable
