+++ NexTrace_Extension/package.json
@@ -11,16 +11,19 @@
     "Other"
   ],
   "activationEvents": [
-    "onCommand:NexTrace.helloWorld",
-    "onView:nextrace-primary-sidebar",
+    "onView:nextrace-primary-sidebar.views",
     "onStartupFinished"
   ],
-  "main": "./react-app/src/extension.js",
+  "main": "./extension.js",
   "contributes": {
     "commands": [
       {
         "command": "NexTrace.startServer",
         "title": "Start Server"
+      },
+      {
+        "command": "myExtension.openWebview",
+        "title": "Open Webview"
       }
     ],
     "viewsContainers": {
@@ -29,28 +32,30 @@
           "id": "nextrace-activitybar",
           "title": "NexTrace",
           "icon": "$(squirrel)",
-          "command": "NexTrace.startServer"
+          "command": "myExtension.openWebview"
         }
       ]
     },
     "views": {
       "nextrace-activitybar": [
         {
-          "id": "nextrace-primary-sidebar",
+          "id": "nextrace-primary-sidebar.views",
           "name": "NexTrace",
-          "icon": "$(squirrel)",
           "contextualTitle": "NexTrace",
-          "type": "webview"
-        }
-      ]
-    }
-  },
+          "type": "webview",
+          "content": "hello world"
+        }]
+    
+      }
+      
+    },
   "scripts": {
     "lint": "eslint .",
     "start": "nodemon /src/server.js",
     "pretest": "npm run lint",
     "test": "node ./test/runTest.js",
-    "build-react": "webpack --config ./react-app/webpack.config.js"
+    "build-react": "webpack --config ./react-app/webpack.config.js",
+    "build-sidepanel": "webpack --config ./react-sidepanel/webpack.config.js"
   },
   "devDependencies": {
     "@babel/core": "^7.23.0",
