+++ NexTrace_Extension/react-app/src/server.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const { request } = require('http');
 const app = express();
-const port = 9999;
+const port = 3695;
 
 app.use(express.json());
