+++ NexTrace_Extension/react-sidepanel/src/components/table.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export default function Table() {
+
+    return (
+        <div className='panel'>
+            <button className='startButton'>Start</button>
+            <p>NexTrace running on port: 3695....</p>
+
+            <button className='buttonOne'></button>
+            <button className='buttonOne'></button>
+            <button className='buttonOne'></button>
+            <button className='buttonOne'></button>
+            <button className='buttonOne'></button>
+
+        </div>
+    )
+}
