+++ NexTrace_Extension/react-sidepanel/webpack.config.js
